feat(pulse): Automatically reconnect to daemon
Just reconnecting doesn't work, we also need to make sure that
exceptions in calls to pulse that happen in between the daemon shutting
down and the adapter reconnecting are properly handled. Otherwise such a
call could crash the whole modules.

So all such calls need to be wrapped in a try-catch block. All calls
from the module into the adapter already do that.